Move vg view to a subcommand
